refactor: destructurization form data types
