+++ ui/form/src/data/verifyRecaptcha.request.ts
@@ -1,22 +1,13 @@
 /** @see https://developers.google.com/recaptcha/docs/verify */
 
-interface ResponseData {
-  success: boolean
-  action?: string
-  score?: number // 0.0 - 1.0
-  challenge_ts?: string
-  hostname?: string
-  'error-codes': string[]
-}
-
-type VerifyRecaptcha = (token: string) => Promise<ResponseData>
+import { VerifyRecaptchaResponse } from './data.interfaces'
 
 const verifyRecaptchaUrl = 'https://www.google.com/recaptcha/api/siteverify'
 
 /** @todo Move to env variables */
 const secretKey = '6LdShlslAAAAAB6HF39ig23TXk2P1lv1UrvyOUlp'
 
-const verifyRecaptcha: VerifyRecaptcha = async (token) =>
+const verifyRecaptcha = async (token: string): Promise<VerifyRecaptchaResponse> =>
   fetch(`${verifyRecaptchaUrl}?secret=${secretKey}&response=${token}`, {
     method: 'post',
   }).then((response) => response.json())
